requirements.txt
aiofiles==23.2.1
anyio==4.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
etelemetry==0.3.1
exceptiongroup==1.1.3
filelock==3.13.1
fitz==0.0.1.dev2
frontend==0.0.3
future==0.18.3
h11==0.14.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
looseversion==1.3.0
lxml==4.9.3
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathlib==1.0.1
pdf2image==1.16.3
Pillow==10.1.0
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
requests==2.31.0
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
traits==6.3.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
